were rewrited some endpoints

+++ src/services/recipe-service.ts
@@ -1,15 +1,16 @@
 import Recipe from "../models/recipe";
 import { recipeData } from "../interfaces/recipeData";
+import { IUserDocument } from "../interfaces/IUserDocument";
 
-const add = async function (data: {
-  name: string;
-}) {
-  const pet = new Recipe(data);
-  await pet.save();
-  return pet;
+const add = async function (data: recipeData, user: IUserDocument) {
+  
+  const recipe = await new Recipe(data);
+  await recipe.save();
+  user.recipes.push(recipe._id);
+  await user.save();
+  return recipe;
+
 };
-
-
 
 const get = async function (id: string) {
   return await Recipe.findById(id);
@@ -19,8 +20,11 @@ const getAll = async function () {
   return await Recipe.find({});
 };
 
-const update = async function (data: recipeData) {
-  return await Recipe.findByIdAndUpdate(data.id, data, { new: true });
+const update = async function (data: recipeData, user: IUserDocument) {
+  if(!user.recipes.includes(data._id)){
+    throw Error("You do not have edit access to this recipe.")
+  }
+  return await Recipe.findByIdAndUpdate(data._id, data, { new: true });
 };
 
 const updateAll = async function (data: recipeData[]) {
@@ -30,8 +34,12 @@ const updateAll = async function (data: recipeData[]) {
 
 };
 
-const del = async function (id: string) {
-  return await Recipe.findByIdAndDelete(id);
+const del = async function (id: string, user: IUserDocument) {
+  if(!user.recipes.includes(id)){
+    throw Error("You do not have permission to delete this recipe.")
+  }
+  await Recipe.findByIdAndDelete(id);
+  return {id}
 };
